Move JavaFX java files under ui directory
JavaFX is considered ui, specifically gui, but the java files arent under the ui directory

Let's move the java files under the ui directory for better organization
