+++ src/main/java/seeyes/Main.java
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import seeyes.ui.MainWindow;
 
 /**
  * TODO: javadocs
